Renamed header files to ".hpp" standard
