+++ realm/wrappers/ExceptionsToManaged.cpp
@@ -1,4 +1,4 @@
-#include "ExceptionsToManaged.h"
+#include "ExceptionsToManaged.hpp"
 
 
 #ifdef WIN32
