+++ realm/wrappers/ExceptionsToManaged.hpp
